Qt: TCP: remove editability of the probe model field
It's fully provided by the cloud API, after all. You can't meaningfully
connect to a probe via TCP without a cert, and the API returns the cert
together with the probe model.
This also means that we need a different way of verifying that the
probe models are actually supported by the running version of Solum.
This is done by simply removing the certification entry for such a
probe from the loaded map. Note that we still preserve all previous and
newly downloaded data in the QSettings file, which allows a mere Solum
upgrade to fix such an issue even without renewing certificates.